+++ test/luacpptemplater_stringhelper.cpp
